Alterações pequenas na classes

+++ FuncaoLogica.java
@@ -2,15 +2,15 @@ import java.util.*;
 
 public class FuncaoLogica {
     private int numeroDeVariaveis;
-    private Set<Integer> mintermos;
-    public FuncaoLogica(int numeroDeVariaveis, Set<Integer> mintermos) {
+    private Set<Long> mintermos;
+    public FuncaoLogica(int numeroDeVariaveis, Set<Long> mintermos) {
         this.numeroDeVariaveis = numeroDeVariaveis;
         this.mintermos = mintermos;
     }
     public int getNumeroDeVariaveis() {
         return numeroDeVariaveis;
     }
-    public Set<Integer> getMintermos() {
+    public Set<Long> getMintermos() {
         return mintermos;
     }
 }
